backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch>=2.1.0
transformer-lens>=2.0.0
sae-lens>=6.27.0
transformers>=4.38.0
safetensors>=0.4.2
httpx>=0.25.0
pydantic>=2.0
python-dotenv>=1.0.0
accelerate>=0.25.0
